Move task class files into task sub package in duke
