+++ src/main/java/duke/TaskList.java
@@ -1,5 +1,10 @@
 package duke;
 
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.Todo;
+
 import java.util.ArrayList;
 import java.util.List;
