+++ src/main/java/duke/task/Event.java
@@ -1,4 +1,6 @@
-package duke;
+package duke.task;
+
+import duke.task.Task;
 
 public class Event extends Task {
     protected String at;
