+++ src/main/java/duke/task/Todo.java
@@ -1,4 +1,6 @@
-package duke;
+package duke.task;
+
+import duke.task.Task;
 
 public class Todo extends Task {
